Add font rendering
This is still pretty janky. I basically added a bitmap font generator
using Java AWT. Then, to render a string I generate a bunch of UIObjects
to the one FontRenderer object, which get passed to the renderer.

If the string is lengthened, it reuses all the UIObjects then adds new ones.
If the string is shortened, it deletes all the excess, and reuses old ones.

It works... So yea.