Fixed normal importing problem, and lighting issues, also added billboards
I had to create a whole stupid parser to parse OBJ files myself,
because for some reason Assimp was importing the normals wrong...
Not sure if I was using it wrong, or blender was exporting it weird
or what, but it works now. Then I experimented with the lighting stuff a bit
to get that ironed out.

I also added billboards, which are just planar icons in world space, that will
always face the camera. These are good for debug icons, now to get the alpha
blending working right...